hw13.1 parallel/nonParallel
