--- hw13.1/src/main/java/ru/sberbank/homework/your_lastname/ParallelThreads.java
@@ -1,15 +1,0 @@
-package ru.sberbank.homework.your_lastname;
-
-import ru.sberbank.homework.common.tasks.CalculationTask;
-import ru.sberbank.homework.common.tasks.SleepyTask;
-import ru.sberbank.homework.common.tasks.StringsTask;
-
-public class ParallelThreads {
-    public static void main(String[] args) {
-        long millis = System.currentTimeMillis();
-        new StringsTask().run();
-        new CalculationTask().run();
-        new SleepyTask().run();
-        System.out.println("All tasks completed! Time: " + (System.currentTimeMillis() - millis));
-    }
-}
